// demos/demo-location/src/lib/api.d.ts
export interface components {
    schemas: {
// ...
        "Circle-Output": {
            /**
             * Areatype
             * @default CIRCLE
             * @constant
             */
            areaType: "CIRCLE";
            center: components["schemas"]["Point"];
            /**
             * Radius
             * @description Expected accuracy for the subscription event of device location, in meters from `center`.
             *     Note: The area surface could be restricted locally depending on regulations. Implementations may enforce a larger minimum radius (e.g. 1000 meters).
             *
             */
            radius: number;
        };
// ...
        VerifyLocationRequest: {
            device?: components["schemas"]["app__schemas__device__Device-Input"] | null;
            area: components["schemas"]["app__schemas__location__Circle"];
            /** Maxage */
            maxAge?: number | null;
        };
// ...
    };
    responses: never;
    parameters: {
        /** @description Correlation id for the different services */
        "x-correlator": string;
    };
    requestBodies: never;
    headers: {
        /** @description Correlation id for the different services */
        "x-correlator": string;
    };
    pathItems: never;
}